torch_overrides: fix masked_scatter dtype mismatch in one-hot matmul
source_2d is bfloat16 (model dtype) but one_hot is float32. Cast
source_2d to float32 for the mm and convert back to data.dtype.

+++ python_package/tt_torch/torch_overrides.py
@@ -78,7 +78,7 @@ class TorchFunctionOverride(TorchFunctionMode):
                 # one-hot via float equality — no int types needed
                 range_n = torch.arange(N, dtype=torch.float32, device=source_2d.device)
                 one_hot = (src_idx_f.reshape(-1, 1) == range_n).float()  # [B*S, N]
-                gathered = torch.mm(one_hot, source_2d).reshape(B, S, H)
+                gathered = torch.mm(one_hot, source_2d.float()).to(data.dtype).reshape(B, S, H)
                 return torch.where(mask, gathered, data)
         return func(*args, **(kwargs or {}))
